+++ packages/crayons-core/src/components/popover/popover.tsx
@@ -9,6 +9,7 @@ import {
   h,
 } from '@stencil/core';
 import { createPopper, Instance } from '@popperjs/core';
+import { PopoverPlacementType, PopoverTriggerType } from '../../utils/types';
 
 @Component({
   tag: 'fw-popover',
@@ -18,7 +19,9 @@ import { createPopper, Instance } from '@popperjs/core';
 export class Popover {
   private popperDiv: HTMLElement;
   private contentRef: HTMLFwListOptionsElement | HTMLElement;
-  private triggerRef: Element;
+  private triggerRef: any;
+  private triggerRefSlot: any = null;
+  private escapeHandler: any = null;
   private overlay: HTMLElement;
 
   @Element() host: HTMLElement;
@@ -59,7 +62,20 @@ export class Popover {
    * The trigger event on which the popover-content is displayed. The available options are
    * 'click' | 'manual' | 'hover', in case of 'manual' no trigger event will be set.
    */
-  @Prop() trigger: 'click' | 'manual' | 'hover' = 'click';
+  @Prop() trigger: PopoverTriggerType = 'click';
+  /**
+   * Option to determine if popover-content has a border.
+   */
+  @Prop() hasBorder = true;
+  /**
+   * Option to prevent the tooltip from being clipped when the component is placed inside a container with
+   * `overflow: auto|hidden|scroll`.
+   */
+  @Prop() hoist = false;
+  /**
+   * Option to disable the popover animation on hide and show.
+   */
+  @Prop() disableTransition = false;
   /**
    * Triggered whenever the popover contents is open/displayed.
    */
@@ -89,7 +105,15 @@ export class Popover {
         ],
       }));
       this.popperInstance.update();
-      this.overlay.style.display = 'block';
+      if (this.trigger !== 'hover') {
+        this.overlay.style.display = 'block';
+      }
+      this.escapeHandler = ((e: any) => {
+        if (e.keyCode === 27) {
+          this.hide();
+        }
+      }).bind(this);
+      document.addEventListener('keydown', this.escapeHandler);
       this.isOpen = !this.isOpen;
       this.fwShow.emit();
     }
@@ -107,7 +131,13 @@ export class Popover {
           { name: 'eventListeners', enabled: false },
         ],
       }));
-      this.overlay.style.display = 'none';
+      if (this.trigger !== 'hover') {
+        this.overlay.style.display = 'none';
+      }
+      if (this.escapeHandler) {
+        document.removeEventListener('keydown', this.escapeHandler);
+        this.escapeHandler = null;
+      }
       this.isOpen = !this.isOpen;
       if (this.contentRef?.tagName === 'FW-LIST-OPTIONS') {
         const listOptionsElement = this.contentRef as HTMLFwListOptionsElement;
@@ -120,7 +150,13 @@ export class Popover {
   componentWillLoad() {
     this.contentRef = this.host.querySelector('[slot="popover-content"]');
     this.triggerRef = this.host.querySelector('[slot="popover-trigger"]');
-    if (this.trigger !== 'manual') {
+    if (this.triggerRef.nodeName === 'SLOT') {
+      const assignedElements = this.triggerRef.assignedElements();
+      if (assignedElements.length) {
+        this.triggerRefSlot = assignedElements[0];
+      }
+    }
+    if (this.trigger === 'click') {
       this.triggerRef.addEventListener(this.trigger, () => {
         if (this.isOpen) {
           this.hide();
@@ -128,9 +164,24 @@ export class Popover {
           this.show();
         }
       });
+    } else if (this.trigger === 'hover') {
+      const trigger = this.triggerRefSlot || this.triggerRef;
+      trigger.addEventListener('focus', this.show.bind(this));
+      trigger.addEventListener('blur', this.hide.bind(this));
+      trigger.addEventListener('mouseenter', this.show.bind(this));
+      trigger.addEventListener('mouseleave', (event: any) => {
+        /**
+         * popovercontent might trigger mouseLeave callback of triggerRef during transition.
+         * This condition avoids flicker that might happen because of the mentioned case.
+         */
+        if (!event.relatedTarget.matches('*[slot="popover-content"]')) {
+          this.hide();
+        }
+      });
     }
     this.popperOptions = {
       placement: this.placement,
+      strategy: this.hoist ? 'fixed' : 'absolute',
       modifiers: [
         {
           name: 'flip',
@@ -159,8 +210,9 @@ export class Popover {
   }
 
   createPopperInstance() {
+    const triggerRef = this.triggerRefSlot || this.triggerRef;
     this.popperInstance = createPopper(
-      this.triggerRef,
+      triggerRef,
       this.popperDiv,
       this.popperOptions
     );
@@ -170,31 +222,23 @@ export class Popover {
     return [
       <slot name='popover-trigger' />,
       <div
-        class='popper-content'
+        class={{
+          'popper-content': true,
+          'no-border': !this.hasBorder,
+          'no-transition': this.disableTransition,
+        }}
         ref={(popperDiv) => (this.popperDiv = popperDiv)}
       >
         <slot name='popover-content' />
       </div>,
-      <div
-        aria-hidden='true'
-        class='overlay'
-        ref={(overlay) => (this.overlay = overlay)}
-        onClick={() => this.hide()}
-      />,
+      this.trigger !== 'hover' && (
+        <div
+          aria-hidden='true'
+          class='overlay'
+          ref={(overlay) => (this.overlay = overlay)}
+          onClick={() => this.hide()}
+        />
+      ),
     ];
   }
 }
-
-export type PopoverPlacementType =
-  | 'top-start'
-  | 'top'
-  | 'top-end'
-  | 'left-start'
-  | 'left'
-  | 'left-end'
-  | 'right-start'
-  | 'right'
-  | 'right-end'
-  | 'bottom-start'
-  | 'bottom'
-  | 'bottom-end';