+++ packages/crayons-core/src/components/radio/radio.e2e.ts
@@ -60,30 +60,21 @@ describe('fw-radio', () => {
     expect(fwSelect.events).toEqual([]);
   });
 
-  it('it should set label when passed as a prop', async () => {
-    const page = await newE2EPage();
-
-    await page.setContent('<fw-radio label="Yes"></fw-radio>');
-    const element = await page.find('fw-radio >>> label');
-    expect(element).toEqualText('Yes');
-  });
-
   it('it should return html structure with slot when content is passed between opening and closing tag', async () => {
     const page = await newE2EPage();
 
-    await page.setContent('<fw-radio label="Yes">Agree</fw-radio>');
+    await page.setContent('<fw-radio description="Yes">Agree</fw-radio>');
     const element = await page.find('fw-radio >>> div');
-    expect(element).toEqualHtml(`<div id="description">
-    <slot/>
+    expect(element).toEqualHtml(`<div class="radio-container">
+    <input type="radio">
+    <label>
+      <span id="label">
+      <slot></slot>
+      </span>
+      <div id="description">
+        Yes
+      </div>
+    </label>
     </div>`);
-  });
-
-  it('it should set aria-label when label is passed as param', async () => {
-    const page = await newE2EPage();
-
-    await page.setContent('<fw-radio value="1" label="yes">1</fw-radio>');
-    const element = await page.find('fw-radio');
-    const label = element.getAttribute('aria-label');
-    expect(label).toBe('yes');
   });
 });